ciao

+++ water-level-monitoring-subsystem/ESP/CV.h
@@ -4,8 +4,8 @@
 
 class CV {
 public:
-    //Help the "NetworkView" to set a new frequence for the periodic sample
-    virtual void setFreq(double newFreq) = 0;
+    //Help the "NetworkView" (NView) to set a new frequence for the periodic sample
+    virtual void setFreq(unsigned long newFreq) = 0;
 };
 
 #endif
